components/backend/requirements.txt
flask==3.1.2
boto3==1.42.46
python-dotenv==1.2.1
sqlalchemy==2.0.46
alembic==1.16.5
psycopg2-binary==2.9.11
faster-whisper==1.2.1
pydantic-settings==2.11.0
